// MotorbikeHome/src/app/model/motos.ts
export interface Motos {
    id_mot: number;
    id_mar: number;
    id_mod: number;
    nom_mot: string;
    img_mot: string;
    cil_mot: string;
    cic_mot: string;
    ref_mot: string;
    trn_mot: string;
    nmar_mot: string;
    cha_mot: string;
    frd_mot: string;
    frt_mot: string;
    rud_mot: string;
    rut_mot: string;
    pes_mot: string;
}